Add API getSelectionRangeInRegion
